feat: split cmd package code (#1932)
